pyproject.toml
[tool.poetry]
name = "bqcli"
version = "0.1.0"
description = ""
authors = ["tmiyamon <miyamototakuya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
prompt_toolkit = "^3.0.4"
google-cloud-bigquery = "^1.24.0"
click = "^7.1.1"
pygments = "^2.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
